Add StaticIPAllocation in Subnet spec and subnetDHCPEnabled in SubnetPort status
1. The NSX operator will allow the customer to specify the property
subnetAdvancedConfig.staticIpAllocation.enabled for VPC Subnet in the Subnet CR.

If staticIPAllocation is not set explicitly in the Subnet CR, the static
IP pool allocation will be automatically enabled when the DHCP mode is DHCPDeactivated
or not set and disabled when the DHCP mode is DHCPServer or DHCPRelay. The NSX operator
will update the spec.subnetAdvancedConfig.staticIpAllocation.enabled with true of false.

The static IP pool allocation and DHCP mode cannot both be enabled simultaneously
in Subnet.

2. The NSX operator will update the subnetDHCPEnabled(true of false) in SubnetPort
status after creating the NSX SubnetPort.

3. The validation of change DHCP mode will be removed. And after the customer changes the
DHCP mode on the Subnet successfully, the NSX operator will update subnetDHCPEnabled on
the status of all the SubnetPorts under the Subnet.

Testing done:
1. create Subnet with the following spec, the NSX Subnet can be created with
subnet_dhcp_config.mode=DHCP_DEACTIVATED and advanced_config.static_ip_allocation.enabled=false.
```
spec:
  accessMode: Private
  advancedConfig:
    staticIPAllocation:
      enabled: false
```
2. create Subnet with the following spec, the NSX Subnet can be created with
subnet_dhcp_config.mode=DHCP_DEACTIVATED and advanced_config.static_ip_allocation.enabled=true.
```
spec:
  accessMode: Private
  advancedConfig:
    staticIPAllocation:
      enabled: true
```
3. create Subnet with the following spec, the NSX Subnet can be created with
.subnet_dhcp_config.mode=DHCP_SERVER and static_ip_allocation.enabled=false.
```
spec:
  accessMode: Private
  subnetDHCPConfig:
    mode: DHCPServer
```
check the Subnet CR, its spec is updated to
```
spec:
  accessMode: Private
  advancedConfig:
    connectivityState: Connected
    enableVLANExtension: false
    staticIPAllocation:
      enabled: false
  ipv4SubnetSize: 32
  subnetDHCPConfig:
    mode: DHCPServer
  vpcName: Project Quality:ns-sunq-01_76261cd3-26a7-4e96-97d5-3886c66f0fa1
```
4. create Subnet with the following spec, the NSX Subnet can be created with
subnet_dhcp_config.mode=DHCP_DEACTIVATED and static_ip_allocation.enabled=true.
```
spec:
  accessMode: Private
```
check the Subnet CR, its spec is updated to
```
spec:
  accessMode: Private
  advancedConfig:
    connectivityState: Connected
    enableVLANExtension: false
    staticIPAllocation:
      enabled: true
  ipv4SubnetSize: 32
  subnetDHCPConfig: {}
  vpcName: Project Quality:ns-sunq-01_76261cd3-26a7-4e96-97d5-3886c66f0fa1
```
5. create Subnet with the following spec
```
spec:
  accessMode: Private
  advancedConfig:
    staticIPAllocation:
      enabled: true
  subnetDHCPConfig:
    mode: DHCPServer
```
the kubectl reports the following error:
The Subnet "subnet-debug-05" is invalid: spec: Invalid value: "object": Static IP pool allocation and Subnet DHCP configuration cannot both be enabled simultaneously in Subnet

6. create SubnetPort on a DHCP Subnet and get the status:
```
status:
  attachment:
    id: df6b9787-794c-5aa4-aaa8-eb71094af9fc
  conditions:
  - lastTransitionTime: "2025-07-07T05:27:08Z"
    message: NSX subnet port has been successfully created/updated
    reason: SubnetPortReady
    status: "True"
    type: Ready
  networkInterfaceConfig:
    ipAddresses:
    - gateway: 172.26.0.1
    logicalSwitchUUID: dc6439fe-f035-4603-a19c-3ab3e46b7a15
    subnetDHCPEnabled: true
```
7. create a SubnetPort on a static IPAM Subnet and get the status:
```
status:
  attachment:
    id: 5e601ac9-aa95-5c48-9ea3-fcd162ceea9d
  conditions:
  - lastTransitionTime: "2025-07-07T05:31:03Z"
    message: NSX subnet port has been successfully created/updated
    reason: SubnetPortReady
    status: "True"
    type: Ready
  networkInterfaceConfig:
    ipAddresses:
    - gateway: 172.26.0.1
      ipAddress: 172.26.0.3/27
    logicalSwitchUUID: 45a8eff3-d054-44b5-b3c3-0a84e6d31370
    macAddress: 04:50:56:00:88:01
```
8. create a SubnetPort on a Subnet with staticIPAllocation disabled and
DHCP disabled, get the status:
```
status:
  attachment:
    id: 8f93fafa-de96-5e74-a7a6-dbd6122f1bc9
  conditions:
  - lastTransitionTime: "2025-07-07T05:34:19Z"
    message: NSX subnet port has been successfully created/updated
    reason: SubnetPortReady
    status: "True"
    type: Ready
  networkInterfaceConfig:
    ipAddresses:
    - gateway: 172.26.0.33
    logicalSwitchUUID: c00a2fb6-5fcd-48a8-b338-e8619acc2f94
```
9. after 6, change the Subnet to disable DHCP, then confirm that subnetDHCPEnabled
is removed from SubnetPort status.
```
status:
  attachment:
    id: df6b9787-794c-5aa4-aaa8-eb71094af9fc
  conditions:
  - lastTransitionTime: "2025-07-07T05:28:29Z"
    message: NSX subnet port has been successfully created/updated
    reason: SubnetPortReady
    status: "True"
    type: Ready
  networkInterfaceConfig:
    ipAddresses:
    - gateway: 172.26.0.1
    logicalSwitchUUID: dc6439fe-f035-4603-a19c-3ab3e46b7a15
```
10. after 7, change the Subnet to enable DHCP, the change fails because of
the error "Static IP pool allocation and Subnet DHCP configuration cannot be enabled simultaneously on a Subnet"
11. after 8, change the Subnet to enable DHCP, then confirm that subnetDHCPEnabled
is added to SubnetPort status.
```
status:
  attachment:
    id: 8f93fafa-de96-5e74-a7a6-dbd6122f1bc9
  conditions:
  - lastTransitionTime: "2025-07-07T05:35:28Z"
    message: NSX subnet port has been successfully created/updated
    reason: SubnetPortReady
    status: "True"
    type: Ready
  networkInterfaceConfig:
    ipAddresses:
    - gateway: 172.26.0.33
    logicalSwitchUUID: c00a2fb6-5fcd-48a8-b338-e8619acc2f94
    subnetDHCPEnabled: true
```

+++ pkg/controllers/subnetport/subnetport_controller.go
@@ -29,6 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
+	"sigs.k8s.io/controller-runtime/pkg/event"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
@@ -164,34 +165,40 @@ func (r *SubnetPortReconciler) Reconcile(ctx context.Context, req ctrl.Request) 
 			r.StatusUpdater.UpdateFail(ctx, subnetPort, err, "Failed to create NSX IPAddressAllocation for AddressBinding restore", setSubnetPortReadyStatusFalse, r.SubnetPortService, r.restoreMode)
 			return common.ResultRequeue, err
 		}
-		nsxSubnetPortState, enableDHCP, err := r.SubnetPortService.CreateOrUpdateSubnetPort(subnetPort, nsxSubnet, "", labels, isVmSubnetPort, r.restoreMode)
+		nsxSubnetPortState, enableDHCP, err := r.SubnetPortService.CreateOrUpdateSubnetPort(subnetPort, nsxSubnet, "", labels, isVmSubnetPort, r.restoreMode, subnetCR)
 		if err != nil {
 			r.StatusUpdater.UpdateFail(ctx, subnetPort, err, "", setSubnetPortReadyStatusFalse, r.SubnetPortService, r.restoreMode)
 			return common.ResultRequeue, err
 		}
-		if nsxSubnetPortState.ExternalAddressBinding == nil && ab == nil {
-			err = r.IpAddressAllocationService.DeleteIPAddressAllocationForAddressBinding(subnetPort)
-			if err != nil {
-				log.Error(err, "Failed to cleanup possible NSX IPAddressAllocation", "SubnetPort", subnetPort)
-				r.StatusUpdater.UpdateFail(ctx, subnetPort, err, "", setSubnetPortReadyStatusFalse, r.SubnetPortService, r.restoreMode)
-				return common.ResultRequeue, err
+		if nsxSubnetPortState == nil {
+			// If nsxSubnetPortState is nil, we only need to update the SubnetDHCPEnabled on the SubnetPort status.
+			subnetPort.Status.NetworkInterfaceConfig.SubnetDHCPEnabled = enableDHCP
+		} else {
+			if nsxSubnetPortState.ExternalAddressBinding == nil && ab == nil {
+				err = r.IpAddressAllocationService.DeleteIPAddressAllocationForAddressBinding(subnetPort)
+				if err != nil {
+					log.Error(err, "Failed to cleanup possible NSX IPAddressAllocation", "SubnetPort", subnetPort)
+					r.StatusUpdater.UpdateFail(ctx, subnetPort, err, "", setSubnetPortReadyStatusFalse, r.SubnetPortService, r.restoreMode)
+					return common.ResultRequeue, err
+				}
 			}
-		}
-		subnetPort.Status.Attachment = v1alpha1.PortAttachment{ID: *nsxSubnetPortState.Attachment.Id}
-		subnetPort.Status.NetworkInterfaceConfig = v1alpha1.NetworkInterfaceConfig{
-			IPAddresses: []v1alpha1.NetworkInterfaceIPAddress{
-				{
-					Gateway: "",
+			subnetPort.Status.Attachment = v1alpha1.PortAttachment{ID: *nsxSubnetPortState.Attachment.Id}
+			subnetPort.Status.NetworkInterfaceConfig = v1alpha1.NetworkInterfaceConfig{
+				IPAddresses: []v1alpha1.NetworkInterfaceIPAddress{
+					{
+						Gateway: "",
+					},
 				},
-			},
-		}
-		if !enableDHCP && len(nsxSubnetPortState.RealizedBindings) > 0 {
-			subnetPort.Status.NetworkInterfaceConfig.IPAddresses[0].IPAddress = *nsxSubnetPortState.RealizedBindings[0].Binding.IpAddress
-			subnetPort.Status.NetworkInterfaceConfig.MACAddress = strings.Trim(*nsxSubnetPortState.RealizedBindings[0].Binding.MacAddress, "\"")
-		}
-		err = r.updateSubnetStatusOnSubnetPort(subnetPort, nsxSubnetPath, nsxSubnet)
-		if err != nil {
-			log.Error(err, "Failed to retrieve Subnet status for SubnetPort", "SubnetPort", subnetPort, "nsxSubnetPath", nsxSubnetPath)
+				SubnetDHCPEnabled: enableDHCP,
+			}
+			if !enableDHCP && len(nsxSubnetPortState.RealizedBindings) > 0 {
+				subnetPort.Status.NetworkInterfaceConfig.IPAddresses[0].IPAddress = *nsxSubnetPortState.RealizedBindings[0].Binding.IpAddress
+				subnetPort.Status.NetworkInterfaceConfig.MACAddress = strings.Trim(*nsxSubnetPortState.RealizedBindings[0].Binding.MacAddress, "\"")
+			}
+			err = r.updateSubnetStatusOnSubnetPort(subnetPort, nsxSubnetPath, nsxSubnet)
+			if err != nil {
+				log.Error(err, "Failed to retrieve Subnet status for SubnetPort", "SubnetPort", subnetPort, "nsxSubnetPath", nsxSubnetPath)
+			}
 		}
 		if reflect.DeepEqual(*old_status, subnetPort.Status) {
 			log.Info("Status (without conditions) already matched", "new status", subnetPort.Status, "existing status", old_status)
@@ -335,8 +342,11 @@ func (r *SubnetPortReconciler) setupWithManager(mgr ctrl.Manager) error {
 			handler.EnqueueRequestsFromMapFunc(r.vmMapFunc),
 			builder.WithPredicates(predicate.LabelChangedPredicate{})).
 		Watches(&v1alpha1.AddressBinding{},
-				handler.EnqueueRequestsFromMapFunc(r.addressBindingMapFunc)).
-		Complete(r) // TODO: watch the virtualmachine event and update the labels on NSX subnet port.
+			handler.EnqueueRequestsFromMapFunc(r.addressBindingMapFunc)).
+		Watches(&v1alpha1.Subnet{},
+			handler.EnqueueRequestsFromMapFunc(r.subnetMapFunc),
+			builder.WithPredicates(OnlyDHCPModeChangedPredicate)).
+		Complete(r)
 }
 
 func (r *SubnetPortReconciler) SetupFieldIndexers(mgr ctrl.Manager) error {
@@ -383,6 +393,56 @@ func (r *SubnetPortReconciler) vmMapFunc(_ context.Context, vm client.Object) []
 		}
 	}
 	return requests
+}
+
+func (r *SubnetPortReconciler) subnetMapFunc(_ context.Context, obj client.Object) []reconcile.Request {
+	subnet, ok := obj.(*v1alpha1.Subnet)
+	if !ok {
+		return nil
+	}
+
+	var ports v1alpha1.SubnetPortList
+	err := r.List(context.Background(), &ports, client.InNamespace(subnet.Namespace))
+	if err != nil {
+		return nil
+	}
+
+	var requests []reconcile.Request
+	for _, sp := range ports.Items {
+		if sp.Spec.Subnet == subnet.Name {
+			requests = append(requests, reconcile.Request{
+				NamespacedName: types.NamespacedName{
+					Name:      sp.Name,
+					Namespace: sp.Namespace,
+				},
+			})
+		}
+	}
+	return requests
+}
+
+var OnlyDHCPModeChangedPredicate = predicate.Funcs{
+	UpdateFunc: func(e event.UpdateEvent) bool {
+		oldSubnet, ok1 := e.ObjectOld.(*v1alpha1.Subnet)
+		newSubnet, ok2 := e.ObjectNew.(*v1alpha1.Subnet)
+		if !ok1 || !ok2 {
+			return false
+		}
+		if oldSubnet.Spec.SubnetDHCPConfig.Mode != newSubnet.Spec.SubnetDHCPConfig.Mode {
+			log.Info("DHCP mode changed for Subnet", "Subnet", newSubnet.Name, "OldMode", oldSubnet.Spec.SubnetDHCPConfig.Mode, "NewMode", newSubnet.Spec.SubnetDHCPConfig.Mode)
+			return true
+		}
+		return false
+	},
+	CreateFunc: func(e event.CreateEvent) bool {
+		return false
+	},
+	DeleteFunc: func(e event.DeleteEvent) bool {
+		return false
+	},
+	GenericFunc: func(e event.GenericEvent) bool {
+		return false
+	},
 }
 
 func (r *SubnetPortReconciler) RestoreReconcile() error {
